requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cycler==0.11.0
flatbuffers==1.12
gast==0.3.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.4
imageai==2.1.6
importlib-metadata==5.1.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
Pillow==7.0.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.4.1
six==1.15.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.11.0
